backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
websockets>=10.0
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
python-dateutil==2.8.2
email-validator==2.1.0.post1
openai>=1.0.0
